+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.7] - 2022-05-15 :gemini:
+- Fixes #257 (bug causing OpenAPI Documentation handler to fail on app start
+  when using PEP 585)
+- Corrects the capitalization of "ApiController" to be "APIController", still
+  keeping the first name for backward compatibility
+- Verifies in tests that `Annotated` is supported by BlackSheep
+
 ## [1.2.6] - 2022-04-28 :candle:
 - Fixes #248, #253
 - Improves support for pydantic (fixes #249)